refactor(desktop): remove the three Tauri commands nothing calls
`team_share_enable_oss`, `team_disconnect_repo` and `team_sync_paths` lost their
last callers when the Team Shared settings section and the Knowledge enable
panel went. Two of them could not work anyway: enabling is a server-side switch
that is already on, and the cloud answers 410 to the disconnect's share-mode
reset.

- `team_share/enable.rs` keeps only the team encryption key and read-only
  status; `enable_oss_impl` and the whole `team_sync_paths` cluster
  (`RegistryWorkspace`, `load_team_workspaces`, the link-collection helpers)
  are gone with their tests
- `team_share/disconnect.rs` is deleted. Its one surviving helper,
  `remove_workspace_team_repo_entry`, moves to `team.rs` beside
  `workspace_delete_team_repo` — its only caller
- the store's `enableOss` / `disconnect` actions go too

Desktop tests: 186 pass / 5 fail, the same HOME-sharing order-sensitive family
that fails on origin/main (5 there as well); each passes run alone.
