+++ apps/desktop/src/commands/team_share/enable.rs
@@ -1,10 +1,9 @@
-//! `enable_oss` / `set_team_secret` / `get_share_status` commands.
+//! `set_team_secret` / `get_team_secret` / `get_share_status` commands.
 //!
-//! These commands wire a team (already created via `team_share::create_team`)
-//! into OSS share mode. They:
-//!   1. Generate (or accept) a 64-char hex team secret and persist it via
-//!      `team_secret_store`.
-//!   2. POST `/v1/teams/{teamId}/share-mode` on FC with `{ mode: "oss" }`.
+//! Enabling is not here any more: share mode is a server-side switch that is
+//! already on for every team, so `enable_oss` had nothing left to decide and
+//! no caller. What remains is the team encryption key (which only the user can
+//! supply) and read-only status.
 //!
 //! The team shared directory is created and linked by the daemon (one global
 //! copy per team under `~/.amuxd/teams/<team_id>/teamclu-team`, exposed via a
@@ -60,143 +59,6 @@ async fn deliver_secrets_and_link(
 }
 
 // ─── Helpers ─────────────────────────────────────────────────────────────
-
-fn generate_team_secret_hex() -> Result<String, String> {
-    let mut bytes = [0u8; 32];
-    getrandom::getrandom(&mut bytes).map_err(|e| format!("getrandom failed: {e}"))?;
-    Ok(hex::encode(bytes))
-}
-
-/// The secret to enable with: caller-supplied > already-stored > freshly minted.
-///
-/// Reading the stored one first is what makes enabling non-destructive. The
-/// callers that used to pass a secret are gone (the three-way wizard's optional
-/// field went with it), so every enable now arrives with `None` — and minting
-/// unconditionally would overwrite a key the user had already configured under
-/// Daemon → General, locking them out of every team value encrypted with it and
-/// encrypting everything after with a key no other member holds.
-///
-/// A store that exists but cannot be read is NOT treated as "no secret": that
-/// is the case where overwriting would destroy something. Only a definite
-/// `NotConfigured` mints a new one.
-fn resolve_team_secret_hex(
-    workspace_path: &str,
-    team_id: &str,
-    team_secret_hex: Option<String>,
-) -> Result<String, String> {
-    if let Some(raw) = team_secret_hex {
-        let normalized = raw.trim().to_ascii_lowercase();
-        if !normalized.is_empty() {
-            if normalized.len() != 64 || !normalized.chars().all(|c| c.is_ascii_hexdigit()) {
-                return Err("team secret must be exactly 64 hex characters".to_string());
-            }
-            return Ok(normalized);
-        }
-    }
-
-    match team_secret_store::load_team_secret(workspace_path, team_id) {
-        Ok(existing) => Ok(existing),
-        Err(team_secret_store::TeamSecretReadError::NotConfigured) => generate_team_secret_hex(),
-        Err(team_secret_store::TeamSecretReadError::StoreUnreadable(e)) => Err(format!(
-            "cannot read the local secret store, so enabling would overwrite a key that may already be there: {e}"
-        )),
-    }
-}
-
-/// The team shared directory is now created and linked by the daemon
-/// (one global copy per team under `~/.amuxd/teams/<team_id>/teamclu-team`,
-/// exposed via a `teamclu-team` symlink in each workspace). Desktop no longer
-/// eagerly creates a per-workspace real directory here; if a real dir is
-/// created later (git clone / OSS engine first write), the daemon consolidates
-/// it into the global copy and replaces it with a symlink. Kept as a no-op so
-/// the enable_* call sites are unchanged and to document the ownership move.
-fn ensure_team_repo_dir(_workspace_path: &str) -> Result<(), String> {
-    Ok(())
-}
-
-async fn post_share_mode(
-    cloud_api_url: &str,
-    team_id: &str,
-    body: &serde_json::Value,
-    access_token: &str,
-) -> Result<(), String> {
-    let fc = FcClient::new(
-        resolve_runtime_fc_endpoint(cloud_api_url)?,
-        access_token.to_string(),
-    );
-    let path = format!("/v1/teams/{}/share-mode", team_id);
-    // The /v1 share-mode endpoint returns 409 once a team's mode is locked.
-    // FcClient::map_fc_response maps any 409 to SyncError::Conflict (its CAS
-    // semantics for /sync/*), which would surface to the UI as the meaningless
-    // "conflict: remote_version=None, remote_cipher_hash=None". Translate it to
-    // a clear, user-facing message instead.
-    fc.post_json(&path, body).await.map_err(|e| match e {
-        SyncError::Conflict { .. } => {
-            "团队共享已开通,无法重复开通或切换共享模式 (share mode already locked)".to_string()
-        }
-        other => other.to_string(),
-    })?;
-    Ok(())
-}
-
-// ─── enable_oss ──────────────────────────────────────────────────────────
-
-pub async fn enable_oss_impl(
-    team_id: String,
-    workspace_path: String,
-    access_token: String,
-    team_secret_hex: Option<String>,
-    cloud_api_url: String,
-) -> Result<EnableShareResult, String> {
-    // Lock the share mode on the server FIRST. If it is already locked (409),
-    // post_share_mode returns a clear error and we bail out BEFORE mutating any
-    // local state — otherwise we would overwrite the existing team secret and
-    // break decryption of data already synced under the original secret.
-    info!(team_id = %team_id, share_mode = "oss", "team_share_enable_oss: locking share-mode on FC");
-    post_share_mode(
-        &cloud_api_url,
-        &team_id,
-        &json!({ "mode": "oss" }),
-        &access_token,
-    )
-    .await?;
-    info!(team_id = %team_id, share_mode = "oss", "team_share_enable_oss: share-mode locked");
-
-    let secret = resolve_team_secret_hex(&workspace_path, &team_id, team_secret_hex)?;
-    team_secret_store::save_team_secret_logged(&workspace_path, &team_id, &secret)?;
-
-    ensure_team_repo_dir(&workspace_path)?;
-
-    // Deliver the OSS team secret to the daemon and link this workspace. The
-    // daemon owns the actual sync/materialization now — desktop no longer
-    // clones or creates the team dir locally. Non-fatal on daemon error.
-    let clone_warning =
-        deliver_secrets_and_link(&team_id, &workspace_path, Some(&secret)).await;
-
-    Ok(EnableShareResult {
-        team_id,
-        share_mode: "oss".to_string(),
-        clone_warning,
-    })
-}
-
-#[tauri::command]
-pub async fn team_share_enable_oss(
-    team_id: String,
-    workspace_path: String,
-    access_token: String,
-    team_secret_hex: Option<String>,
-    cloud_api_url: String,
-) -> Result<EnableShareResult, String> {
-    enable_oss_impl(
-        team_id,
-        workspace_path,
-        access_token,
-        team_secret_hex,
-        cloud_api_url,
-    )
-    .await
-}
 
 // ─── set_team_secret ────────────────────────────────────────────────────
 
@@ -331,231 +193,6 @@ pub async fn team_share_get_status(
     get_share_status_impl(team_id, workspace_path, access_token, cloud_api_url).await
 }
 
-// ─── team_sync_paths ───────────────────────────────────────────────────────
-//
-// Surfaces, for the settings UI, *where team content physically lives* and
-// *every `teamclu-team` symlink that points at it*. All three share modes
-// (oss / managed_git / custom_git) converge on the same topology: one real
-// directory per team at `~/.amuxd/teams/<team_id>/teamclu-team` (the daemon's
-// global copy; git modes clone into it), with a `teamclu-team` symlink in each
-// workspace that joined the team. The list of workspaces comes from the Cloud
-// API (`GET /v1/workspaces?teamId=...`, the single source of truth for a
-// team's workspaces), filtered to paths that exist on this machine.
-
-/// One workspace's `teamclu-team` entry, for the settings UI.
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkspaceLinkInfo {
-    pub workspace_path: String,
-    pub display_name: String,
-    /// `<workspace_path>/teamclu-team` — the symlink (or legacy local dir).
-    pub link_path: String,
-    /// `"symlink"` | `"real_dir"` | `"missing"` (see `detect_link_status`).
-    pub status: String,
-    /// Whether this is the workspace currently open in the app.
-    pub is_current: bool,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TeamSyncPaths {
-    /// `~/.amuxd/teams/<team_id>/teamclu-team` — the single real copy.
-    pub real_dir: Option<String>,
-    /// Whether that real directory currently exists on disk.
-    pub real_dir_exists: bool,
-    /// The current workspace plus every workspace bound to this team in the
-    /// daemon registry, each with its `teamclu-team` link status.
-    pub links: Vec<WorkspaceLinkInfo>,
-}
-
-/// One workspace as returned by the Cloud API — kept minimal (only the
-/// fields the settings UI's symlink materialization actually needs).
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct RegistryWorkspace {
-    #[serde(default)]
-    pub(crate) path: String,
-    #[serde(default)]
-    pub(crate) display_name: String,
-    #[serde(default)]
-    pub(crate) team_id: Option<String>,
-}
-
-/// Raw item shape of `GET /v1/workspaces?teamId=...` — `name` is the display
-/// name, `slug` carries the filesystem path (mirrors the daemon's
-/// `WorkspaceRow` mapping in `apps/daemon/src/backend/cloud_api/mod.rs`).
-#[derive(Debug, Default, Deserialize)]
-struct CloudWorkspaceItem {
-    #[serde(default)]
-    name: String,
-    #[serde(default)]
-    path: Option<String>,
-    #[serde(default)]
-    slug: Option<String>,
-    #[serde(default, rename = "teamId")]
-    team_id: Option<String>,
-}
-
-#[derive(Debug, Default, Deserialize)]
-struct CloudWorkspacePage {
-    #[serde(default)]
-    items: Vec<CloudWorkspaceItem>,
-}
-
-/// Keep only cloud workspaces whose path exists as a directory on *this*
-/// machine. Pure (no network) so it's easy to unit-test in isolation — the
-/// cloud `workspaces` table intentionally includes other devices' workspace
-/// paths, which this daemon/desktop must never touch. Mirrors the daemon's
-/// `cloud_rows_to_local_linkable_paths` filter (Task 8) for consistency.
-fn filter_workspaces_present_on_disk(items: Vec<RegistryWorkspace>) -> Vec<RegistryWorkspace> {
-    items
-        .into_iter()
-        .filter(|w| !w.path.is_empty() && std::path::Path::new(&w.path).is_dir())
-        .collect()
-}
-
-/// Fetch `team_id`'s workspaces from the Cloud API and filter to paths that
-/// exist locally. Best-effort: any network/parse/auth failure yields an empty
-/// list rather than surfacing an error, matching the previous (toml-read)
-/// behavior's fail-open semantics for this settings-UI-only surface.
-pub(crate) async fn load_team_workspaces(
-    team_id: &str,
-    _workspace_path: &str,
-    access_token: &str,
-    cloud_api_url: &str,
-) -> Vec<RegistryWorkspace> {
-    if team_id.trim().is_empty()
-        || access_token.trim().is_empty()
-        || cloud_api_url.trim().is_empty()
-    {
-        return vec![];
-    }
-    let endpoint = match resolve_runtime_fc_endpoint(cloud_api_url) {
-        Ok(endpoint) => endpoint,
-        Err(_) => return vec![],
-    };
-    let fc = FcClient::new(endpoint, access_token.to_string());
-    let path = format!(
-        "/v1/workspaces?teamId={}&limit=200",
-        urlencoding::encode(team_id)
-    );
-    let items = match fc.get_json(&path).await {
-        Ok(value) => serde_json::from_value::<CloudWorkspacePage>(value)
-            .map(|p| p.items)
-            .unwrap_or_default(),
-        Err(_) => return vec![],
-    };
-    let workspaces: Vec<RegistryWorkspace> = items
-        .into_iter()
-        .filter_map(|item| {
-            let path = item.path.or(item.slug)?;
-            if path.trim().is_empty() {
-                return None;
-            }
-            Some(RegistryWorkspace {
-                path,
-                display_name: item.name,
-                team_id: item.team_id.or_else(|| Some(team_id.to_string())),
-            })
-        })
-        .collect();
-    filter_workspaces_present_on_disk(workspaces)
-}
-
-/// Best-effort canonicalization for stable path comparison; falls back to the
-/// raw string when the path can't be canonicalized (e.g. it doesn't exist).
-fn canon_path(path: &str) -> String {
-    std::fs::canonicalize(path)
-        .map(|p| p.to_string_lossy().into_owned())
-        .unwrap_or_else(|_| path.to_string())
-}
-
-fn basename_or(path: &str) -> String {
-    std::path::Path::new(path)
-        .file_name()
-        .map(|n| n.to_string_lossy().into_owned())
-        .unwrap_or_else(|| path.to_string())
-}
-
-fn push_workspace_link(
-    links: &mut Vec<WorkspaceLinkInfo>,
-    seen: &mut std::collections::HashSet<String>,
-    ws_path: &str,
-    display_name: String,
-    is_current: bool,
-) {
-    if ws_path.is_empty() || !seen.insert(canon_path(ws_path)) {
-        return;
-    }
-    let link_path = std::path::Path::new(ws_path)
-        .join(TEAM_REPO_DIR)
-        .to_string_lossy()
-        .into_owned();
-    links.push(WorkspaceLinkInfo {
-        workspace_path: ws_path.to_string(),
-        display_name,
-        link_path,
-        status: detect_link_status(ws_path).to_string(),
-        is_current,
-    });
-}
-
-pub async fn team_sync_paths_impl(
-    team_id: String,
-    workspace_path: String,
-    access_token: Option<String>,
-    cloud_api_url: Option<String>,
-) -> TeamSyncPaths {
-    let real_dir = global_team_dir_display(&team_id);
-    let real_dir_exists = real_dir
-        .as_deref()
-        .map(|p| std::path::Path::new(p).exists())
-        .unwrap_or(false);
-
-    let current_canon = canon_path(&workspace_path);
-    let mut links: Vec<WorkspaceLinkInfo> = Vec::new();
-    let mut seen: std::collections::HashSet<String> = std::collections::HashSet::new();
-
-    // Current workspace first, even if it isn't team-bound in the registry yet
-    // (e.g. just enabled, daemon hasn't recorded team_id). Dedup handles the
-    // overlap when it *is* already in the registry.
-    if !workspace_path.is_empty() {
-        let name = basename_or(&workspace_path);
-        push_workspace_link(&mut links, &mut seen, &workspace_path, name, true);
-    }
-
-    // Other team workspaces come from the Cloud API and require an auth
-    // token; when the caller doesn't have one handy (best-effort settings-UI
-    // surface), we still show the current workspace above and simply skip
-    // the rest rather than failing the whole command.
-    if let (Some(token), Some(cloud_api_url)) = (access_token, cloud_api_url) {
-        for w in load_team_workspaces(&team_id, &workspace_path, &token, &cloud_api_url).await {
-            let is_current = canon_path(&w.path) == current_canon;
-            let name = if w.display_name.is_empty() {
-                basename_or(&w.path)
-            } else {
-                w.display_name.clone()
-            };
-            push_workspace_link(&mut links, &mut seen, &w.path, name, is_current);
-        }
-    }
-
-    TeamSyncPaths {
-        real_dir,
-        real_dir_exists,
-        links,
-    }
-}
-
-#[tauri::command]
-pub async fn team_sync_paths(
-    team_id: String,
-    workspace_path: String,
-    access_token: Option<String>,
-    cloud_api_url: Option<String>,
-) -> Result<TeamSyncPaths, String> {
-    Ok(team_sync_paths_impl(team_id, workspace_path, access_token, cloud_api_url).await)
-}
-
 #[cfg(test)]
 mod link_status_tests {
     use super::*;
@@ -591,69 +228,4 @@ mod link_status_tests {
         }
     }
 
-    #[test]
-    fn filter_workspaces_present_on_disk_keeps_only_existing_dirs() {
-        let ws = tempfile::tempdir().unwrap();
-        let existing = ws.path().to_str().unwrap().to_string();
-        let missing = ws
-            .path()
-            .join("does-not-exist")
-            .to_string_lossy()
-            .into_owned();
-
-        let items = vec![
-            RegistryWorkspace {
-                path: existing.clone(),
-                display_name: "alpha".into(),
-                team_id: Some("team-1".into()),
-            },
-            RegistryWorkspace {
-                path: missing,
-                display_name: "beta".into(),
-                team_id: Some("team-1".into()),
-            },
-            RegistryWorkspace {
-                path: String::new(),
-                display_name: "gamma".into(),
-                team_id: Some("team-1".into()),
-            },
-        ];
-
-        let got = filter_workspaces_present_on_disk(items);
-        assert_eq!(got.len(), 1);
-        assert_eq!(got[0].path, existing);
-    }
-
-    #[tokio::test]
-    async fn load_team_workspaces_short_circuits_without_team_or_token() {
-        // No network involved: empty team_id / access_token both bail before
-        // any HTTP call is attempted, so this is safe to run in CI.
-        assert!(
-            load_team_workspaces("", "/tmp", "token", "https://example.test")
-                .await
-                .is_empty()
-        );
-        assert!(
-            load_team_workspaces("team-1", "/tmp", "", "https://example.test")
-                .await
-                .is_empty()
-        );
-    }
-
-    #[tokio::test]
-    async fn sync_paths_includes_current_workspace_with_link_status() {
-        let ws = tempfile::tempdir().unwrap();
-        let ws_path = ws.path().to_str().unwrap().to_string();
-        // No access token → the cloud fetch is skipped entirely (best-effort),
-        // so this stays a pure local-disk test like the pre-refactor version.
-        let out = team_sync_paths_impl("team-xyz".into(), ws_path.clone(), None, None).await;
-
-        // Current workspace is always present and flagged, even when it isn't
-        // team-bound in the cloud list yet. With no teamclu-team entry on
-        // disk → "missing".
-        let current: Vec<_> = out.links.iter().filter(|l| l.is_current).collect();
-        assert_eq!(current.len(), 1);
-        assert_eq!(current[0].status, "missing");
-        assert!(current[0].link_path.ends_with(TEAM_REPO_DIR));
-    }
 }
