+++ apps/desktop/src/lib.rs
@@ -479,7 +479,6 @@ pub fn run() {
             commands::terminal::terminal_write,
             commands::team::workspace_read_team_meta,
             commands::team::workspace_delete_team_repo,
-            commands::team_share::disconnect::team_disconnect_repo,
             commands::team_sync_proxy::team_sync_repo,
             commands::team_sync_proxy::team_file_versions,
             commands::team_sync_proxy::team_file_content,
@@ -575,11 +574,9 @@ pub fn run() {
             commands::tray_menu::update_tray_menu_labels,
             commands::app_menu::update_app_menu_labels,
             commands::team_share::team_share_create,
-            commands::team_share::enable::team_share_enable_oss,
             commands::team_share::enable::team_share_set_team_secret,
             commands::team_share::enable::team_share_get_team_secret,
             commands::team_share::enable::team_share_get_status,
-            commands::team_share::enable::team_sync_paths,
             commands::team_share::join::team_share_join_existing,
             commands::team_litellm::team_litellm_setup,
             commands::team_sync_proxy::oss_sync_now,
