+++ apps/desktop/tests/team_share_smoke.rs
@@ -135,99 +135,6 @@ async fn create_team_slim_only_calls_v1_teams_and_returns_id_slug() {
 // ─── Task 6 tests ─────────────────────────────────────────────────────────
 
 #[tokio::test]
-async fn enable_oss_provisions_secret_and_local_dir() {
-    let _guard = HOME_GUARD.lock().unwrap_or_else(|e| e.into_inner());
-    let server = MockServer::start().await;
-    Mock::given(method("POST"))
-        .and(path("/v1/teams/t1/share-mode"))
-        .and(header("authorization", "Bearer test-jwt"))
-        .and(body_partial_json(json!({ "mode": "oss" })))
-        .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-            "id": "t1",
-            "share_mode": "oss",
-        })))
-        .mount(&server)
-        .await;
-
-    let tmp = TempDir::new().expect("tempdir");
-    isolate_home(&tmp);
-    let workspace = seed_workspace(&tmp, &server.uri());
-
-    let result = team_share::enable::enable_oss_impl(
-        "t1".to_string(),
-        workspace.clone(),
-        "test-jwt".to_string(),
-        None,
-        server.uri(),
-    )
-    .await
-    .expect("enable_oss should succeed");
-    assert_eq!(result.team_id, "t1");
-    assert_eq!(result.share_mode, "oss");
-
-    // Secret persisted (64 lowercase hex chars).
-    let secret = team_secret_store::load_team_secret(&workspace, "t1")
-        .expect("secret should be readable after enable_oss");
-    assert_eq!(secret.len(), 64, "team secret must be 64 hex chars");
-    assert!(
-        secret.chars().all(|c| c.is_ascii_hexdigit()),
-        "secret should be hex"
-    );
-
-    // teamclu-team is created+linked by the daemon, not by enable_oss.
-    let team_repo_dir = std::path::Path::new(&workspace).join("teamclu-team");
-    assert!(
-        !team_repo_dir.exists()
-            || team_repo_dir
-                .symlink_metadata()
-                .unwrap()
-                .file_type()
-                .is_symlink(),
-        "enable must not create a real teamclu-team dir (daemon owns it)"
-    );
-
-    // NOTE: team fields (oss_team_id / share_mode) are intentionally NOT
-    // persisted to teamclu.json anymore — the single source of truth is the
-    // current-team store (commits ad563711 / b1baec40). The returned
-    // share_mode / team_id are asserted above.
-}
-
-#[tokio::test]
-async fn enable_oss_uses_provided_team_secret() {
-    let _guard = HOME_GUARD.lock().unwrap_or_else(|e| e.into_inner());
-    let server = MockServer::start().await;
-    Mock::given(method("POST"))
-        .and(path("/v1/teams/t1/share-mode"))
-        .and(header("authorization", "Bearer test-jwt"))
-        .and(body_partial_json(json!({ "mode": "oss" })))
-        .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-            "id": "t1",
-            "share_mode": "oss",
-        })))
-        .mount(&server)
-        .await;
-
-    let tmp = TempDir::new().expect("tempdir");
-    isolate_home(&tmp);
-    let workspace = seed_workspace(&tmp, &server.uri());
-
-    let provided = "abcdef0123456789".repeat(4);
-    team_share::enable::enable_oss_impl(
-        "t1".to_string(),
-        workspace.clone(),
-        "test-jwt".to_string(),
-        Some(provided.to_uppercase()),
-        server.uri(),
-    )
-    .await
-    .expect("enable_oss with provided secret should succeed");
-
-    let loaded =
-        team_secret_store::load_team_secret(&workspace, "t1").expect("secret should be readable");
-    assert_eq!(loaded, provided);
-}
-
-#[tokio::test]
 async fn set_team_secret_validates_and_stores() {
     let _guard = HOME_GUARD.lock().unwrap_or_else(|e| e.into_inner());
     let tmp = TempDir::new().expect("tempdir");
